\name{protein}
\title{Data Frame 'protein'}
\alias{protein}
\description{
This is the Protein Data Set from the UCI Machine Learning Repository.
It contains information about protein concentration in different samples.
} 
\usage{
protein
} 
\format{
A data frame with
45730
rows and
10
columns.
\itemize{
\item{\code{SampleID}: A unique identifier for each sample.}
\item{\code{Protein1}: Concentration of Protein 1.}
\item{\code{Protein2}: Concentration of Protein 2.}
\item{\code{Protein3}: Concentration of Protein 3.}
\item{\code{Protein4}: Concentration of Protein 4.}
\item{\code{Protein5}: Concentration of Protein 5.}
\item{\code{Protein6}: Concentration of Protein 6.}
\item{\code{Protein7}: Concentration of Protein 7.}
\item{\code{Protein8}: Concentration of Protein 8.}
\item{\code{Protein9}: Concentration of Protein 9.}
\item{\code{Protein10}: Concentration of Protein 10.}
} 
} 
